+++ src/main/java/gov/cms/madie/models/measure/GroupScoringPopulation.java
@@ -1,0 +1,9 @@
+package gov.cms.madie.models.measure;
+
+import java.util.Map;
+
+public interface GroupScoringPopulation {
+  String getScoring();
+
+  Map<MeasurePopulation, ?> getPopulation();
+}
