+++ src/main/java/gov/cms/madie/models/measure/HapiOperationOutcome.java
@@ -1,0 +1,17 @@
+package gov.cms.madie.models.measure;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class HapiOperationOutcome {
+  private int code;
+  private String message;
+  // Plain object as we don't know or care about the structure of the HAPI FHIR response
+  private Object outcomeResponse;
+}
