+++ src/main/java/gov/cms/madie/models/measure/MeasurePopulationOption.java
@@ -1,0 +1,13 @@
+package gov.cms.madie.models.measure;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class MeasurePopulationOption {
+  MeasurePopulation measurePopulation;
+  boolean required;
+}
