+++ src/main/java/gov/cms/madie/models/measure/TestCasePopulationValue.java
@@ -1,0 +1,17 @@
+package gov.cms.madie.models.measure;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TestCasePopulationValue {
+
+  private MeasurePopulation name;
+  private Boolean expected;
+  private Boolean actual;
+}
