+++ src/main/java/gov/cms/madie/models/validators/EnumValidator.java
@@ -1,0 +1,28 @@
+package gov.cms.madie.models.validators;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import javax.validation.ReportAsSingleViolation;
+import javax.validation.constraints.NotNull;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@NotNull(message = "Value cannot be null")
+@ReportAsSingleViolation
+@Target({ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Constraint(validatedBy = EnumValidatorImpl.class)
+public @interface EnumValidator {
+
+  Class<? extends Enum<?>> enumClass();
+
+  String message() default "Value provided is not a valid option.";
+
+  Class<?>[] groups() default {};
+
+  Class<? extends Payload>[] payload() default {};
+}
