+++ src/main/java/gov/cms/madie/models/validators/EnumValidatorImpl.java
@@ -1,0 +1,25 @@
+package gov.cms.madie.models.validators;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+public class EnumValidatorImpl implements ConstraintValidator<EnumValidator, String> {
+
+  private List<String> acceptedValues;
+
+  @Override
+  public void initialize(EnumValidator annotation) {
+    acceptedValues =
+      Stream.of(annotation.enumClass().getEnumConstants())
+        .map(Enum::toString)
+        .collect(Collectors.toList());
+  }
+
+  @Override
+  public boolean isValid(String value, ConstraintValidatorContext context) {
+    return acceptedValues.contains(value);
+  }
+}
