--- src/main/java/gov/cms/madiejavamodels/measure/GroupScoringPopulation.java
@@ -1,9 +1,0 @@
-package gov.cms.madiejavamodels.measure;
-
-import java.util.Map;
-
-public interface GroupScoringPopulation {
-  String getScoring();
-
-  Map<MeasurePopulation, ?> getPopulation();
-}
