--- src/main/java/gov/cms/madiejavamodels/measure/TestCasePopulationValue.java
@@ -1,17 +1,0 @@
-package gov.cms.madiejavamodels.measure;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TestCasePopulationValue {
-
-  private MeasurePopulation name;
-  private Boolean expected;
-  private Boolean actual;
-}
