--- src/main/java/gov/cms/madiejavamodels/utils/ScoringPopulationDefinition.java
@@ -1,35 +1,0 @@
-package gov.cms.madiejavamodels.utils;
-
-import gov.cms.madiejavamodels.measure.MeasurePopulation;
-import gov.cms.madiejavamodels.measure.MeasurePopulationOption;
-import gov.cms.madiejavamodels.measure.MeasureScoring;
-
-import java.util.List;
-import java.util.Map;
-
-public interface ScoringPopulationDefinition {
-  Map<MeasureScoring, List<MeasurePopulationOption>> SCORING_POPULATION_MAP =
-    Map.of(
-      MeasureScoring.RATIO,
-      List.of(
-        new MeasurePopulationOption(MeasurePopulation.INITIAL_POPULATION, true),
-        new MeasurePopulationOption(MeasurePopulation.NUMERATOR, true),
-        new MeasurePopulationOption(MeasurePopulation.NUMERATOR_EXCLUSION, false),
-        new MeasurePopulationOption(MeasurePopulation.DENOMINATOR, true),
-        new MeasurePopulationOption(MeasurePopulation.DENOMINATOR_EXCLUSION, false)),
-      MeasureScoring.PROPORTION,
-      List.of(
-        new MeasurePopulationOption(MeasurePopulation.INITIAL_POPULATION, true),
-        new MeasurePopulationOption(MeasurePopulation.NUMERATOR, true),
-        new MeasurePopulationOption(MeasurePopulation.NUMERATOR_EXCLUSION, false),
-        new MeasurePopulationOption(MeasurePopulation.DENOMINATOR, true),
-        new MeasurePopulationOption(MeasurePopulation.DENOMINATOR_EXCLUSION, false),
-        new MeasurePopulationOption(MeasurePopulation.DENOMINATOR_EXCEPTION, false)),
-      MeasureScoring.CONTINUOUS_VARIABLE,
-      List.of(
-        new MeasurePopulationOption(MeasurePopulation.INITIAL_POPULATION, true),
-        new MeasurePopulationOption(MeasurePopulation.MEASURE_POPULATION, true),
-        new MeasurePopulationOption(MeasurePopulation.MEASURE_POPULATION_EXCLUSION, false)),
-      MeasureScoring.COHORT,
-      List.of(new MeasurePopulationOption(MeasurePopulation.INITIAL_POPULATION, true)));
-}
